requirements.txt
slack-bolt>=1.21.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
